poc_producer/requirements.txt
faststream[redis]==0.5.0
redis==5.0.1
uvicorn==0.24.0
pydantic==2.6.0
fast-depends==2.4.12
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-instrumentation-asyncio>=0.40b0
opentelemetry-instrumentation-redis>=0.40b0